Health regions yearly population estimates per age group and sex

This function binds UF male and female datasets (uf_male_pop and uf_female_pop), aggregates estimatives by health regions and includes a new variable called 'sex'.

Usage

regsaude_sex_pop(type = "standard", source = "datasus")

Arguments

type

character. 'standard' or 'reg_saude_449'

source

character. datasus for Brazilian Health Ministry estimates, or ufrn for UFRN-DEM-LEPP estimates.

Value

A tibble.

See also